requirements.txt
Flask==2.1.1
gunicorn==20.1.0
twilio==7.5.0
google-cloud-aiplatform==1.8.1
google-cloud-storage==1.42.3
Pillow==9.0.1
requests==2.27.1
